refactor: Move remaining pure Kotlin tests to shared modules
Moved to domain/commonTest:
- LoadingResultTest — tests domain model
- DoorFcmModelTest — tests FCM topic name generation
- FcmTopicTest — FCM topic contract tests

Moved to data/commonTest:
- FcmPayloadParsingTest — tests FcmPayloadParser

Deleted from androidApp:
- ServerConfigRepositoryTest — duplicate of CachedServerConfigRepositoryTest

androidApp now has only 7 test files, all genuinely Android-specific:
AppConfigFactory, Room schema (2), network fixtures, UI tests (3).
