+++ AndroidGarage/domain/src/commonTest/kotlin/com/chriscartland/garage/domain/model/DoorFcmModelTest.kt
@@ -1,25 +1,7 @@
-/*
- * Copyright 2024 Chris Cartland. All rights reserved.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- *
- */
+package com.chriscartland.garage.domain.model
 
-package com.chriscartland.garage.fcm
-
-import com.chriscartland.garage.domain.model.toFcmTopic
-import org.junit.Assert.assertEquals
-import org.junit.Test
+import kotlin.test.Test
+import kotlin.test.assertEquals
 
 class DoorFcmModelTest {
     @Test
@@ -54,7 +36,6 @@ class DoorFcmModelTest {
 
     @Test
     fun typicalBuildTimestamp() {
-        // Colons and slashes should be replaced, hyphens and dots preserved
         val topic = "2024-01-15T10:30:00Z".toFcmTopic()
         assertEquals("door_open-2024-01-15T10.30.00Z", topic.string)
     }
