+++ AndroidGarage/domain/src/commonTest/kotlin/com/chriscartland/garage/domain/model/LoadingResultTest.kt
@@ -1,29 +1,9 @@
-/*
- * Copyright 2024 Chris Cartland. All rights reserved.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- *
- */
+package com.chriscartland.garage.domain.model
 
-package com.chriscartland.garage.door
-
-import com.chriscartland.garage.domain.model.DoorEvent
-import com.chriscartland.garage.domain.model.DoorPosition
-import com.chriscartland.garage.domain.model.LoadingResult
-import org.junit.Assert.assertEquals
-import org.junit.Assert.assertNull
-import org.junit.Assert.assertTrue
-import org.junit.Test
+import kotlin.test.Test
+import kotlin.test.assertEquals
+import kotlin.test.assertIs
+import kotlin.test.assertNull
 
 class LoadingResultTest {
     @Test
@@ -66,19 +46,19 @@ class LoadingResultTest {
     @Test
     fun loadingIsCorrectType() {
         val result: LoadingResult<String> = LoadingResult.Loading("data")
-        assertTrue(result is LoadingResult.Loading)
+        assertIs<LoadingResult.Loading<String>>(result)
     }
 
     @Test
     fun completeIsCorrectType() {
         val result: LoadingResult<String> = LoadingResult.Complete("data")
-        assertTrue(result is LoadingResult.Complete)
+        assertIs<LoadingResult.Complete<String>>(result)
     }
 
     @Test
     fun errorIsCorrectType() {
         val result: LoadingResult<String> = LoadingResult.Error(RuntimeException())
-        assertTrue(result is LoadingResult.Error)
+        assertIs<LoadingResult.Error>(result)
     }
 
     @Test
